Move UWP PipePlatform
